+++ addons/iap/static/src/js/insufficient_credit_error_handler.js
@@ -1,7 +1,7 @@
 /** @odoo-module */
 import { Dialog } from "@web/core/dialog/dialog";
 import { registry } from "@web/core/registry";
-import { useService } from "@web/core/service_hook";
+import { useService } from "@web/core/utils/hooks";
 
 class InsufficientCreditDialog extends Dialog {
     setup() {
