+++ addons/im_livechat/__manifest__.py
@@ -99,9 +99,8 @@ Help your customers with this chat, and analyse their feedback.
             'web/static/src/core/browser/browser.js',
             'web/static/src/core/browser/feature_detection.js',
             'web/static/src/core/dialog/dialog.js',
-            'web/static/src/core/effect_hook.js',
             'web/static/src/core/errors/error_dialogs.js',
-            'web/static/src/core/hotkey_hook.js',
+            'web/static/src/core/utils/hooks.js',
             'web/static/src/core/hotkeys/hotkey_service.js',
             'web/static/src/core/l10n/dates.js',
             'web/static/src/core/l10n/localization.js',
@@ -113,7 +112,6 @@ Help your customers with this chat, and analyse their feedback.
             'web/static/src/core/notifications/notification_container.js',
             'web/static/src/core/notifications/notification_service.js',
             'web/static/src/core/registry.js',
-            'web/static/src/core/service_hook.js',
             'web/static/src/core/ui/block_ui.js',
             'web/static/src/core/ui/ui_service.js',
             'web/static/src/core/user_service.js',
