+++ addons/web/static/src/legacy/legacy_dialog.js
@@ -2,10 +2,8 @@
 
 import { Dialog } from "../core/dialog/dialog";
 import { patch } from "@web/core/utils/patch";
-import { useEffect } from "@web/core/effect_hook";
+import { useEffect } from "@web/core/utils/hooks";
 import OwlDialog from "web.OwlDialog";
-
-const { hooks } = owl;
 
 /**
  * This is a patch of the new Dialog class.
