+++ addons/web/static/src/webclient/actions/action_dialog.js
@@ -1,9 +1,9 @@
 /** @odoo-module **/
 
-import { Dialog } from "../../core/dialog/dialog";
-import { DebugMenu } from "../../core/debug/debug_menu";
-import { useOwnDebugContext } from "../../core/debug/debug_context";
-import { useEffect } from "../../core/effect_hook";
+import { Dialog } from "@web/core/dialog/dialog";
+import { DebugMenu } from "@web/core/debug/debug_menu";
+import { useOwnDebugContext } from "@web/core/debug/debug_context";
+import { useEffect } from "@web/core/utils/hooks";
 
 const { hooks } = owl;
