+++ addons/web/static/src/webclient/navbar/navbar.js
@@ -1,12 +1,11 @@
 /** @odoo-module **/
 
-import { Dropdown } from "../../core/dropdown/dropdown";
-import { DropdownItem } from "../../core/dropdown/dropdown_item";
-import { useService } from "../../core/service_hook";
-import { useEffect } from "../../core/effect_hook";
-import { registry } from "../../core/registry";
-import { debounce } from "../../core/utils/timing";
-import { ErrorHandler, NotUpdatable } from "../../core/utils/components";
+import { Dropdown } from "@web/core/dropdown/dropdown";
+import { DropdownItem } from "@web/core/dropdown/dropdown_item";
+import { useEffect, useService } from "@web/core/utils/hooks";
+import { registry } from "@web/core/registry";
+import { debounce } from "@web/core/utils/timing";
+import { ErrorHandler, NotUpdatable } from "@web/core/utils/components";
 
 const { Component, hooks } = owl;
 const { useExternalListener, useRef } = hooks;
